fix(dfn): mark transient blocks (#178)
The iper variable was previously missed. It has in_record true but it's not in a record, as per mf6's definition, it's in the period block. I suppose in_record really means 'is a child of a composite', which is in tension with the hard distinction mf6 makes between blocks and variables. In the TOML format we no longer need in_record.

IMO it's cleaner to just mark transient blocks as such instead of defining an extra index variable. This does mean that we will need to translate the transient flag back into 'iper' terminology for the mf6io guide.