refactor(models): distinguish source repo via prefix (#202)
Add a prefix to model names identifying which repository the model is from (i.e., what "kind" of model it is):

- example/... example models from https://github.com/MODFLOW-ORG/modflow6-examples
- test/... test models from https://github.com/MODFLOW-ORG/modflow6-testmodels (only mf6)
- large/... models from https://github.com/MODFLOW-ORG/modflow6-largetestmodels

This will allow more easily selecting subsets of models by the repository they are from as we do in MF6 autotests.

I considered adding an "mf6" prefix to all these (useful if we want to expose older models, e.g. from the mf5to6 folder in the test models repo). I also considered making the prefix identical to the repository name. Not sure which way is best, this will do for now. Anyone reading this with an opinion, please make it known.

Also, remove the auto-generated model functions (doesn't play nice with "namespaced" i.e. "/"-delimited model names), tweak the registry introspection functions for consistency, and update some docstrings.