refactor(models): introduce model registry classes (#210)
Introduce a ModelRegistry base class with LocalRegistry and PoochRegistry variants. Addresses an open topic from #134:

    One question is whether the last two use cases can be merged somehow, i.e. how to give the models API the ability to access local models. It is often necessary to modify a local model during development, and it is inconvenient to have to switch between the models API and the fixtures for this.

The aim here is for an mf6 developer to be able to transparently swap in a local registry in order to access local models, where otherwise the "official" registry would be the (user- and developer-facing) default.

This changeset should not break any existing usages of the preexisting model registry, fingers crossed.