refactor(LocalRegistry): rework index method prefix parameters (#215)
The parameter prefix is a string to prepend to model names. But the routine to search for model dirs also has a prefix param to filter the results. Switch the prefix parameter's semantics to match this, and pass it through, and add a new model_name_prefix parameter for model name modification.